refactor: merge ui-components into components folder
