+++ screens/SignUp/BirthDateScreen.tsx
@@ -6,7 +6,7 @@ import {
   View
 } from 'react-native';
 
-import CicularButton from '../../ui-components/CircularButton';
+import CircularButton from '../../components/CircularButton';
 
 
 
@@ -15,7 +15,7 @@ const BirthDateScreen = () => {
   return (
     <View>
       <Text style={styles.text}>When were you born?</Text>
-      <ArrowButton title='>' />
+      <CircularButton title='>' />
     </View>
   );
 }
